backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pandas>=1.5.0
numpy>=1.23.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
polygon-api-client>=1.14.0
motor>=3.3.0
pymongo>=4.6.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
